Move converters and definition to base facet
