+++ pint/facets/group/definitions.py
@@ -12,9 +12,10 @@ import re
 from dataclasses import dataclass
 from typing import Tuple
 
-from ...definitions import Definition, UnitDefinition
+from ...definitions import Definition
 from ...errors import DefinitionSyntaxError
 from ...util import SourceIterator
+from ..base.definitions import UnitDefinition
 
 
 @dataclass(frozen=True)
